Add files via upload
The main technology in play here is Bootstrap 4. The cdn version is used here because there has been some issues. 
Using an offline version of the bootstrap docs by devdocs, I am being able to understand the workflow of Bootstrap.
it is being slow and taking a lot of time and effort from me. But progress is being made.
It is obviously easier and better to setup a bootstrap development environment than using the cdn.